Removed console.log statements and changed date format
Removed console.log statements from the grouping algorithm in the master
branch.  Console.log statements will still be used in development
branches.  Also changed the date format used in the "Build date:"
section so that it is D-MMM-YYYY and is consistent with the date format
that is used in the rest of jsGroup and its GitHub repository